moved legacy pages from views folder to pages folder
