+++ Xamarin/SSW.Rewards/SSW.Rewards/Pages/LoginPage.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              NavigationPage.HasNavigationBar="False"
-             x:Class="SSW.Rewards.Views.LoginPage">
+             x:Class="SSW.Rewards.Pages.LoginPage">
     <ContentPage.Content>
         <Grid>
             <Image Source="rewardsbackground"
