+++ Xamarin/SSW.Rewards/SSW.Rewards/Pages/TestPushPage.xaml
@@ -3,7 +3,7 @@
     xmlns="http://xamarin.com/schemas/2014/forms"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     NavigationPage.HasNavigationBar="False"
-    x:Class="SSW.Rewards.Views.TestPushPage">
+    x:Class="SSW.Rewards.Pages.TestPushPage">
     <ContentPage.Content>
         <Grid>
             <Image Source="login_background"
