+++ Xamarin/SSW.Rewards/SSW.Rewards/ViewModels/OnBoardingViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using SSW.Rewards.Services;
-using SSW.Rewards.Views;
+using SSW.Rewards.Pages;
 using System;
 using System.Collections.ObjectModel;
 using System.Threading.Tasks;
