--- Xamarin/SSW.Rewards/SSW.Rewards/Views/EarnPointsPage.xaml
@@ -1,143 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:controls="clr-namespace:SSW.Rewards.Controls"
-             mc:Ignorable="d"
-             Title="{Binding Title}"
-			 x:Name="EarnPointsView"
-             x:Class="SSW.Rewards.Views.EarnPointsPage">
-    <Shell.TitleView>
-        <controls:TitleView Title="{Binding Title}" />
-    </Shell.TitleView>
-    <ContentPage.Content>
-        <Grid Padding="0">
-            <Image Aspect="AspectFill"
-                   Source="polygon"
-                   RelativeLayout.WidthConstraint="{ConstraintExpression Type=RelativeToParent, Property=Width}"
-                   RelativeLayout.HeightConstraint="{ConstraintExpression Type=RelativeToParent, Property=Height}"/>
-
-            <AbsoluteLayout>
-
-                <CollectionView AbsoluteLayout.LayoutBounds="0,0,1,1"
-                                AbsoluteLayout.LayoutFlags="All"
-                                Margin="10"
-                                ItemsSource="{Binding ExternalRewards}">
-                    <CollectionView.ItemsLayout>
-                        <GridItemsLayout Orientation="Vertical"
-                                         Span="2"/>
-                    </CollectionView.ItemsLayout>
-                    <d:CollectionView.ItemsSource>
-                        <x:Array Type="{x:Type x:String}">
-                            <x:String>Link Twitter</x:String>
-                            <x:String>Watch Presentations</x:String>
-                            <x:String>Take Quiz</x:String>
-                        </x:Array>
-                    </d:CollectionView.ItemsSource>
-                    <CollectionView.ItemTemplate>
-                        <DataTemplate>
-                            <ContentView>
-                                <Frame WidthRequest="100"
-                                       HeightRequest="220"
-                                       CornerRadius="20"
-									   x:Name="Item"
-                                       Margin="10"
-                                       Padding="0"
-                                       BackgroundColor="{StaticResource LeaderboardBackground}"
-                                       HasShadow="True">
-                                    <AbsoluteLayout Padding="0"
-													InputTransparent="True">                                      
-                                        <Image Source="{Binding Picture}"
-											   InputTransparent="True"
-                                                AbsoluteLayout.LayoutBounds="0,0,1,0.6"
-                                                AbsoluteLayout.LayoutFlags="All"/>
-
-										<Image Source="points_background"
-											   InputTransparent="True"
-                                               AbsoluteLayout.LayoutBounds="1,0.52,0.6,0.5"
-                                               IsVisible="{Binding IsBonus}"
-                                               Opacity="0.4"
-                                               AbsoluteLayout.LayoutFlags="All">
-                                            <Image.TranslationY>
-                                                <OnPlatform x:TypeArguments="x:Double">
-                                                    <On Platform="Android">-10</On>
-                                                </OnPlatform>
-                                            </Image.TranslationY>
-                                        </Image>
-
-										<Image Source="icon_goldstar"
-											   InputTransparent="True"
-                                               IsVisible="{Binding IsBonus}"
-                                               AbsoluteLayout.LayoutBounds="0.5,0.51,0.15,0.15"
-                                               AbsoluteLayout.LayoutFlags="All">
-                                            <Image.TranslationY>
-                                                <OnPlatform x:TypeArguments="x:Double">
-                                                    <On Platform="Android">-10</On>
-                                                </OnPlatform>
-                                            </Image.TranslationY>
-                                        </Image>
-
-                                        <Label Text="{Binding Points}"
-                                               TextColor="{StaticResource PointsColour}"
-                                               IsVisible="{Binding IsBonus}"
-											   InputTransparent="True"
-                                               FontAttributes="Bold"
-                                               FontSize="22"
-                                               VerticalTextAlignment="Center"
-                                               AbsoluteLayout.LayoutBounds="0.995,0.51,0.4,0.2"
-                                               AbsoluteLayout.LayoutFlags="All">
-                                            <Label.TranslationY>
-                                                <OnPlatform x:TypeArguments="x:Double">
-                                                    <On Platform="Android">-10</On>
-                                                </OnPlatform>
-                                            </Label.TranslationY>
-                                        </Label>
-
-										<Label Text="{Binding Title}"
-											   InputTransparent="True"
-                                               VerticalOptions="End"
-                                               TextColor="Black"
-                                               Margin="5,10"
-                                               AbsoluteLayout.LayoutBounds="0.1,0.8,0.8,0.3"
-                                               AbsoluteLayout.LayoutFlags="All"/>
-                                    </AbsoluteLayout>
-
-									<Frame.GestureRecognizers>
-										<TapGestureRecognizer BindingContext="{Binding Source={x:Reference EarnPointsView}, Path=BindingContext}"
-                                                              Command="{Binding OnEventTapped}"
-                                                              CommandParameter="{Binding  Source={x:Reference Item}, Path=BindingContext.Url}"
-                                                              NumberOfTapsRequired="1" />
-									</Frame.GestureRecognizers>
-                                </Frame>
-                            </ContentView>
-                        </DataTemplate>
-                    </CollectionView.ItemTemplate>
-                </CollectionView>
-
-                <BoxView AbsoluteLayout.LayoutBounds="1,1,1,0.115"
-                         AbsoluteLayout.LayoutFlags="All"
-                         HeightRequest="250"
-                         Opacity="0.5"
-                         Color="Black"/>
-
-                <Image AbsoluteLayout.LayoutBounds="0.5,0.95,90,90"
-                       AbsoluteLayout.LayoutFlags="PositionProportional"
-                       Source="camera">
-                    <Image.GestureRecognizers>
-                        <TapGestureRecognizer NumberOfTapsRequired="1"
-                                              Command="{Binding OnScanTapped}"/>
-                    </Image.GestureRecognizers>
-                </Image>
-
-                <Label AbsoluteLayout.LayoutBounds="0,0.99,1,0.04"
-                       HorizontalTextAlignment="Center"
-                       AbsoluteLayout.LayoutFlags="All"
-                       Text="Scan QR Code"
-                       TextColor="White"/>
-
-            </AbsoluteLayout>
-
-        </Grid>
-    </ContentPage.Content>
-</ContentPage>
