--- Xamarin/SSW.Rewards/SSW.Rewards/Views/LeaderSummary.xaml
@@ -1,77 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-<ViewCell xmlns="http://xamarin.com/schemas/2014/forms" 
-          xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-          xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-          xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-          xmlns:ffimage="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
-          xmlns:transforms="clr-namespace:FFImageLoading.Transformations;assembly=FFImageLoading.Transformations"
-          xmlns:converters="clr-namespace:SSW.Rewards.Converters"
-          mc:Ignorable="d"
-          x:Class="SSW.Rewards.Views.LeaderSummary">
-  <ViewCell.View>
-
-        <StackLayout Orientation="Horizontal"
-                     HorizontalOptions="FillAndExpand"
-                     BackgroundColor="{Binding IsMe, Converter={converters:IsMeToColorConverter}}">
-
-            <BoxView HeightRequest="100"
-                     WidthRequest="5"
-                     BackgroundColor="Red"
-                     IsVisible="{Binding IsMe}"
-                     d:IsVisible="true"/>
-
-            <Grid HorizontalOptions="FillAndExpand"
-                  VerticalOptions="Center"
-                  Padding="5">
-
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="1*"/>
-                    <ColumnDefinition Width="3*"/>
-                    <ColumnDefinition Width="4*"/>
-                    <ColumnDefinition Width="3*"/>
-                    <ColumnDefinition Width="3*"/>
-                </Grid.ColumnDefinitions>
-
-                <Label Style="{StaticResource Leaderboard}"
-                       Grid.Column="0"
-                       FontSize="18"
-                       Text="{Binding Rank}"
-                       VerticalOptions="Center"
-                       Margin="5,0,0,0"
-                       d:Text="2"/>
-
-                <ffimage:CachedImage Source="{Binding ProfilePic}"
-                                     Grid.Column="1"
-                                     LoadingPlaceholder="icon_avatar"
-                                     ErrorPlaceholder="icon_avatar"
-                                     d:Source="MattG"
-                                     Aspect="AspectFill"
-                                     CacheType="Disk"
-                                     CacheDuration="5"
-                                     VerticalOptions="Center"
-                                     WidthRequest="40"
-                                     Margin="10,0,5,0">
-                    <ffimage:CachedImage.Transformations>
-                        <transforms:CircleTransformation BorderSize="5"
-                                                         BorderHexColor="{StaticResource ProfileBorder}"/>
-                    </ffimage:CachedImage.Transformations>
-                </ffimage:CachedImage>
-
-                <Label Style="{StaticResource Leaderboard}"
-                       Grid.Column="2"
-                       Text="{Binding Name}"
-                       VerticalOptions="Center"
-                       d:Text="Matt Goldman"/>
-
-                <Label Style="{StaticResource Leaderboard}"
-                       Grid.Column="3"
-                       Text="{Binding BaseScore, StringFormat='{0} points'}"
-                       d:Text="120pts"
-                       VerticalOptions="Center"
-                       HorizontalOptions="End"
-                       Margin="0,0,5,0"/>
-
-                </Grid>
-        </StackLayout>
-  </ViewCell.View>
-</ViewCell>
