--- Xamarin/SSW.Rewards/SSW.Rewards/Views/LeaderSummary.xaml.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SSW.Rewards.ViewModels;
-using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-
-namespace SSW.Rewards.Views
-{
-    [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class LeaderSummary : ViewCell
-    {
-        public LeaderSummary()
-        {
-
-        }
-
-        public LeaderSummary(LeaderViewModel viewModel)
-        {
-            InitializeComponent();
-            BindingContext = viewModel;
-        }
-    }
-}
