--- Xamarin/SSW.Rewards/SSW.Rewards/Views/RewardsPage.xaml
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ContentPage xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:controls="clr-namespace:SSW.Rewards.Controls"
-             xmlns:d="http://xamarin.com/schemas/2014/forms/design"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             mc:Ignorable="d"
-             xmlns:ffimage="clr-namespace:FFImageLoading.Forms;assembly=FFImageLoading.Forms"
-             x:Name="RewardPage"
-             x:Class="SSW.Rewards.Views.RewardsPage">
-    <Shell.TitleView>
-        <controls:TitleView x:Name="pageTitle" Title="{Binding Title}"/>
-    </Shell.TitleView>
-    <Grid Padding="0">
-        <Image Aspect="AspectFill"
-                   Source="polygon"
-                   RelativeLayout.WidthConstraint="{ConstraintExpression Type=RelativeToParent, Property=Width}"
-                   RelativeLayout.HeightConstraint="{ConstraintExpression Type=RelativeToParent, Property=Height}"/>
-
-        <AbsoluteLayout>
-            <CollectionView AbsoluteLayout.LayoutBounds="0,0,1,1"
-                            AbsoluteLayout.LayoutFlags="All"
-                            Margin="10"
-                            ItemsSource="{Binding Rewards}">
-                <d:CollectionView.ItemsSource>
-                    <x:Array Type="{x:Type x:String}">
-                        <x:String>SSW Smart Keepcup</x:String>
-                        <x:String>Xaiomi Mi Band</x:String>
-                        <x:String>Azure Superpowers Ticket</x:String>
-                    </x:Array>
-                </d:CollectionView.ItemsSource>
-                <CollectionView.ItemTemplate>
-                    <DataTemplate>
-                        <ContentView>
-                            <Frame CornerRadius="0"
-                                   BackgroundColor="Transparent"
-                                   HeightRequest="100"
-                                   HasShadow="True"
-                                   Padding="10"
-                                   Margin="0,5"
-                                   x:Name="Item">
-                                <Frame.Background>
-                                    <LinearGradientBrush EndPoint="1,0"
-                                                         StartPoint="0,1">
-                                        <GradientStop Color="Red"
-                                                      Offset="0.1"/>
-                                        <GradientStop Color="Transparent"
-                                                      Offset="1.0"/>
-                                    </LinearGradientBrush>
-                                </Frame.Background>
-                                <Frame.GestureRecognizers>
-                                    <TapGestureRecognizer Command="{Binding RewardCardTappedCommand}"
-                                                          BindingContext="{Binding Source={x:Reference RewardPage}, Path=BindingContext}"
-                                                          CommandParameter="{Binding Source={x:Reference Item}, Path=BindingContext}"
-                                                          NumberOfTapsRequired="1"/>
-                                </Frame.GestureRecognizers>
-                                <Grid ColumnDefinitions="5*, 5*"
-                                      HorizontalOptions="FillAndExpand"
-                                      VerticalOptions="FillAndExpand"
-                                      Padding="0"
-                                      RowDefinitions="5*, 5*"
-                                      InputTransparent="True">
-                                    <!-- The below ffimage is the actual image used in the app -->
-                                    <!-- This should be uncommebted before deploying -->
-                                    
-                                    <ffimage:CachedImage Grid.Column="0"
-                                                         Grid.RowSpan="2"
-                                                         Source="{Binding ImageUri}"
-                                                         LoadingPlaceholder="dev_loading_placeholder"
-                                                         InputTransparent="True"/>
-                                    
-                                    <!-- The below image placeholder is for use in the XAML previewer -->
-                                    <!-- This should be commented out or removed before deploying -->
-                                    <!--<Image Grid.Column="0"
-                                           Grid.RowSpan="2"
-                                           Source="prize_keepcup"
-                                           />-->
-                                    
-                                    <Label Grid.Column="1"
-                                           Grid.RowSpan="2"
-                                           Text="{Binding Name}"
-                                           TextColor="White"
-                                           FontAttributes="Bold"
-                                           FontSize="22"
-                                           d:Text="SSW smart keepcup"
-                                           InputTransparent="True"/>
-
-                                    <Button BackgroundColor="Black"
-                                            TextColor="Red"
-                                            HorizontalOptions="EndAndExpand"
-                                            VerticalOptions="EndAndExpand"
-                                            HeightRequest="40"
-                                            WidthRequest="70"
-                                            Margin="0"
-                                            Text="More"
-                                            Command="{Binding MoreTapped}"
-                                            BindingContext="{Binding Source={x:Reference RewardPage}, Path=BindingContext}"
-                                            CommandParameter="{Binding Source={x:Reference Item}, Path=BindingContext}"
-                                            Grid.Row="1"
-                                            Grid.Column="1"/>
-
-                                </Grid>
-                            </Frame>
-                        </ContentView>
-                    </DataTemplate>
-                </CollectionView.ItemTemplate>
-            </CollectionView>
-        </AbsoluteLayout>
-    </Grid>
-</ContentPage>
